Move chatter model item to a separate file
